refactor: 승패 판단 책임을 승패에게 이동
Result의 Map이 LinkedHashMap일 필요 없어서 변경
